Update README.md
This is a "rat in a maze" one of the famous problem in data structures.I have used java swing . This approach solves the problem using recursion. I have used the concept of Depth First Search. A rat can move in any of eight directions from its current position. After runnig the program in the terminal , it asks for you to enter the starting and ending positions of the maze.And it will show you the frame where a grid form (to represent a maze) will be shown . There you can click any cell to make walls. Once you created the walls,you can go to the menu button in that window and press start. You can view how the problem is getting solved in each step!